+++ client/lib/createOptDis.js
@@ -17,17 +17,18 @@ function enrich(type, payload) {
 function createOptDis(schema, dispatch, restClient, serverActionQueue) {
     const plurals = lodash_1.default.keys(schema).map(k => pluralize_1.default(k));
     function enqueueServerAction(action) {
-        serverActionQueue.addItem(action);
+        return serverActionQueue.addItem(action);
     }
     function dispatchServerAction() {
-        const action = serverActionQueue.getItem();
-        if (!action)
+        const item = serverActionQueue.getItem();
+        if (!item)
             return;
+        const { id, action } = item;
         restClient.post('/jason/api/action', action)
-            .then(serverActionQueue.itemProcessed)
-            .catch(e => {
-            dispatch({ type: 'upsertLocalUi', data: { error: JSON.stringify(e) } });
-            serverActionQueue.itemProcessed();
+            .then(({ data }) => serverActionQueue.itemProcessed(id, data))
+            .catch(error => {
+            dispatch({ type: 'jason/upsert', payload: { error } });
+            serverActionQueue.itemFailed(id, error);
         });
     }
     setInterval(dispatchServerAction, 10);
@@ -36,7 +37,7 @@ function createOptDis(schema, dispatch, restClient, serverActionQueue) {
         const data = enrich(type, payload);
         dispatch({ type, payload: data });
         if (plurals.indexOf(type.split('/')[0]) > -1) {
-            enqueueServerAction({ type, payload: data });
+            return enqueueServerAction({ type, payload: data });
         }
     };
 }